enable support for eth beacon single method tests
- rename EthConsensus to EthBeacon for related classes
- enable support for eth beacon methods to be tested in single method test mode
- moved method classes to methods module
- fixed incorrect EthValidatorStatus, removed PENDING_ONGOING as it doesn't exist, added PENDING_QUEUED
- enable retry for fetching blocks from eth beacon endpoint
-
