+++ chainbench/profile/evm/all.py
@@ -4,10 +4,10 @@
 
 from locust import constant_pacing
 
-from chainbench.user import EvmMethods
-from chainbench.util.cli import get_subclass_methods
+from chainbench.user.methods import EvmMethods
+from chainbench.user.methods.common import get_subclass_functions
 
 
 class EvmAllProfile(EvmMethods):
     wait_time = constant_pacing(1)
-    tasks = [EvmMethods.get_method(method) for method in get_subclass_methods(EvmMethods)]
+    tasks = [task.method for task in get_subclass_functions(EvmMethods)]
