+++ chainbench/test_data/ethereum.py
@@ -4,6 +4,8 @@ import typing as t
 from argparse import Namespace
 from dataclasses import dataclass
 
+from tenacity import retry, stop_after_attempt
+
 from chainbench.test_data.blockchain import (
     Block,
     BlockNotFoundError,
@@ -11,6 +13,7 @@ from chainbench.test_data.blockchain import (
     BlockRange,
     TestData,
 )
+from chainbench.util.http import HttpClient, HttpErrorLevel, HttpStatusError
 from chainbench.util.rng import RNG
 
 logger = logging.getLogger(__name__)
@@ -20,15 +23,15 @@ Epoch = int
 
 class EthValidatorStatus:
     ACTIVE = "active"
-    ACTIVE_ONGOING = "active_ongoing"
     ACTIVE_EXITING = "active_exiting"
+    ACTIVE_ONGOING = "active_ongoing"
     ACTIVE_SLASHED = "active_slashed"
-    PENDING = "pending"
-    PENDING_ONGOING = "pending_ongoing"
-    PENDING_INITIALIZED = "pending_initialized"
     EXITED = "exited"
-    EXITED_UNSLASHED = "exited_unslashed"
     EXITED_SLASHED = "exited_slashed"
+    EXITED_UNSLASHED = "exited_unslashed"
+    PENDING = "pending"
+    PENDING_INITIALIZED = "pending_initialized"
+    PENDING_QUEUED = "pending_queued"
     WITHDRAWAL = "withdrawal"
     WITHDRAWAL_DONE = "withdrawal_done"
     WITHDRAWAL_POSSIBLE = "withdrawal_possible"
@@ -55,7 +58,7 @@ class EthCommittee:
 
 
 @dataclass(frozen=True)
-class EthConsensusBlock(Block):
+class EthBeaconBlock(Block):
     epoch: int
     committees: list[EthCommittee]
 
@@ -77,7 +80,7 @@ class EthConsensusBlock(Block):
             for validator in committee["validators"]:
                 validators.append(EthValidator(validator))
             committees.append(EthCommittee(committee["index"], validators))
-        return EthConsensusBlock(slot, epoch, committees)
+        return EthBeaconBlock(slot, epoch, committees)
 
     def get_random_committee(self, rng: RNG) -> EthCommittee:
         return rng.random.choice(self.committees)
@@ -94,14 +97,22 @@ class EthConsensusBlock(Block):
         return [validator.index for validator in committee.validators[start_index : start_index + count]]
 
 
-class EthConsensusTestData(TestData[EthConsensusBlock]):
+class EthBeaconTestData(TestData[EthBeaconBlock]):
+    def init_http_client(self, host_url: str):
+        self._client = HttpClient(host_url, error_level=HttpErrorLevel.ServerError)
+        self._logger.debug("Host: %s", host_url)
+
     def fetch_block_header(self, block_id: int | str) -> dict[str, t.Any]:
         block_response = self.client.get(f"/eth/v1/beacon/headers/{block_id}")
         if block_response.status_code == 404:
             raise BlockNotFoundError
+        elif block_response.status_code != 200:
+            raise HttpStatusError(
+                block_response.status_code, f"Failed to fetch block header: {block_response.status_message}"
+            )
         return block_response.json["data"]["header"]["message"]
 
-    def fetch_block(self, block_id: int | str) -> EthConsensusBlock:
+    def fetch_block(self, block_id: int | str) -> EthBeaconBlock:
         if isinstance(block_id, str):
             if (block_id := block_id.lower()) not in (
                 "head",
@@ -109,14 +120,19 @@ class EthConsensusTestData(TestData[EthConsensusBlock]):
                 "finalized",
             ):
                 raise ValueError("Invalid block identifier")
-        slot = int(self.fetch_block_header(block_id)["slot"])
+        try:
+            slot = int(self.fetch_block_header(block_id)["slot"])
 
-        committees_response = self.client.get(f"/eth/v1/beacon/states/{block_id}/committees", params={"slot": slot})
-        return EthConsensusBlock.from_response(slot, committees_response.json)
+            committees_response = self.client.get(f"/eth/v1/beacon/states/{block_id}/committees", params={"slot": slot})
+            return EthBeaconBlock.from_response(slot, committees_response.json)
+        except BlockNotFoundError:
+            return self.fetch_latest_block()
 
-    def fetch_latest_block(self) -> EthConsensusBlock:
+    @retry(reraise=True, stop=stop_after_attempt(5))
+    def fetch_latest_block(self) -> EthBeaconBlock:
         return self.fetch_block("head")
 
+    @retry(reraise=True, stop=stop_after_attempt(5))
     def fetch_latest_block_number(self) -> BlockNumber:
         return int(self.fetch_block_header("head")["slot"])
 
@@ -129,7 +145,7 @@ class EthConsensusTestData(TestData[EthConsensusBlock]):
         logger.info("Using blocks from %s to %s as test data", start_block_number, end_block_number)
         return BlockRange(start_block_number, end_block_number)
 
-    def get_block_from_data(self, data: dict[str, t.Any] | str) -> EthConsensusBlock:
+    def get_block_from_data(self, data: dict[str, t.Any] | str) -> EthBeaconBlock:
         def get_committee(committee: dict[str, t.Any]) -> EthCommittee:
             return EthCommittee(
                 index=committee["index"],
@@ -144,7 +160,7 @@ class EthConsensusTestData(TestData[EthConsensusBlock]):
         slot = data_dict["block_number"]
         epoch = slot // 32
         committees = [get_committee(committee) for committee in data_dict["committees"]]
-        return EthConsensusBlock(slot, epoch, committees)
+        return EthBeaconBlock(slot, epoch, committees)
 
     def get_random_epoch(self, rng: RNG) -> Epoch:
         return self.get_random_block(rng).epoch
